refactor: Split config package so it doesn't end up importing core, core/vm etc.
