+++ plugin/evm/admin.go
@@ -10,6 +10,7 @@ import (
 	"github.com/ava-labs/avalanchego/api"
 	"github.com/ava-labs/avalanchego/utils/profiler"
 	"github.com/ava-labs/coreth/plugin/evm/client"
+	"github.com/ava-labs/coreth/plugin/evm/config"
 	"github.com/ethereum/go-ethereum/log"
 )
 
@@ -79,6 +80,6 @@ func (p *Admin) SetLogLevel(_ *http.Request, args *client.SetLogLevelArgs, reply
 }
 
 func (p *Admin) GetVMConfig(_ *http.Request, _ *struct{}, reply *client.ConfigReply) error {
-	reply.Config = &p.vm.config
+	reply.Config = (*config.Config)(&p.vm.config)
 	return nil
 }
